LMS-2994 (1/2) - Use self-made SELECT
The code which checks for activity completion needs to retrieve the data
from the database. Change the SELECT statement used there from one
created by the database library functions to a manually generated one.

So far, there is no change in functionality. This commit just simplifies
the next changes to the code.