LMS-2994 tests (4/8) - Use hacked fixture class
The class ensures that we always work with non-preview output during
testing. This is necessary to receive a real date that has been read
from the database instead of being given a preview (example) date.

We pass a pre-calculated value for the name property during test setup
so that we will not produce a string lookup error. The latter, though
harmless, is very annoying.