LMS-2994, follow-up - only accept activity completion if not failed
Adjust meaning of activity completion in case of activities that can be
passed or failed, like assignments: do not include entries in the
completion table element if the student has failed the activity. It is
still included if the activity is passed or completion is not combined
with a pass/fail marking.

This commit adjusts the non-test code to behave in the expected way
described above.