Work around old setting for PHP_CodeChecker.
Our current Travis setup pulls in moodle-plugin-ci in version 1. The
setting for the rule which codechecker errors out on
(Squiz.WhiteSpace.OperatorSpacing) has been configured differently
during recent updates to moodle-local_codechecker, which is enabled by
moodle-plugin-ci, version 2.

We are tracking the proper fix in LMS-3013, but for now, this will do.